requirements.txt
langchain 
wikipedia
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic